+++ lib/rubocop/cop/katalyst_cops.rb
@@ -1,3 +1,4 @@
 # frozen_string_literal: true
 
+require_relative "koi/duplicates_association"
 require_relative "koi/table_link_heading"